WebQuanLi/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
httpx>=0.24.1
jinja2>=3.1.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
